Make a basic test for PacketFinder.
I haven't run them on the BBB yet. I plan to make a few more
tests, perhaps for different sized packets.

I also got rid of a macro that I thought was unecessary.
(Brian might beg to differ.)

+++ bbb_cape/src/bbb/uart_reader.cc
@@ -109,7 +109,7 @@ UartReader::~UartReader() {
   if (fd_ > 0) close(fd_);
 }
 
-int UartReader::ReadBytes(AlignedChar *dest, size_t max_bytes) {
+ssize_t UartReader::ReadBytes(AlignedChar *dest, size_t max_bytes) {
   return read(fd_, dest, max_bytes);
 }
